silenced more compiler warnings on amd64

+++ branches/vermont/dtls/concentrator/AutoFocus.cpp
@@ -334,7 +334,7 @@ void AutoFocus::metalist()
 			sprintf(num,"%25s",(*iter)->global.c_str());
 			locl.append(num);
 			locl.append("\t");
-			sprintf(num,"%10llu",data);
+			sprintf(num,"%10llu",(long long unsigned int) data);
 			locl.append(num);
 			locl.append(" \t");
 			percentage = (double) (data*100) / (double) (*iter)->numTotal;
